Finish part 4 blog list auth and tests!

+++ part4/bloglist/controllers/blogs.js
@@ -1,8 +1,10 @@
 const blogsRouter = require('express').Router();
-const Blog = require('../models//blog');
+const jwt = require('jsonwebtoken');
+const Blog = require('../models/blog');
+const User = require('../models/user');
 
 blogsRouter.get('/', async (req, res) => {
-  const blogs = await Blog.find({});
+  const blogs = await Blog.find({}).populate('user');
   res.json(blogs);
 });
 
@@ -21,8 +23,12 @@ blogsRouter.get('/:id', async (req, res, next) => {
 
 blogsRouter.post('/', async (req, res, next) => {
   try {
-    const blog = new Blog(req.body);
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    const user = await User.findById(decodedToken.id);
+    const blog = new Blog({ ...req.body, user: user._id });
     const savedBlog = await blog.save();
+    user.blogs = user.blogs.concat(savedBlog._id);
+    await user.save();
     res.status(201).json(savedBlog);
   } catch (err) {
     next(err);
@@ -31,6 +37,14 @@ blogsRouter.post('/', async (req, res, next) => {
 
 blogsRouter.delete('/:id', async (req, res, next) => {
   try {
+    const decodedToken = jwt.verify(req.token, process.env.SECRET);
+    const blogToDelete = await Blog.findById(req.params.id);
+    const user = await User.findById(decodedToken.id);
+    if (blogToDelete.user.toString() !== user._id.toString()) {
+      return res.status(401).json({
+        error: 'deletion permitted by creator only'
+      });
+    }
     await Blog.findByIdAndDelete(req.params.id);
     res.status(204).end();
   } catch (err) {
